# backend/back-neovook/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/article/create": {
      "post": {
        "operationId": "ArticleController_createArticle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Article"
        ]
      }
    },
...
  },
  "info": {
    "title": "Api neovook",
    "description": "Enssembles des routes de neovook",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Article": {
        "type": "object",
        "properties": {}
      }
    }
  }
}